#298 Split common setup from Controller and put in Application
